Added a utility to safely copy one array to another.
This accounts for integer overflow/underflow, out of range, nulls,
special values, etc. when the destination array does not support the
representations of the source array. Users can choose to throw an
error or insert a placeholder value (with or without a warning).

This setter is now used inside the convertToWasmArray functions.